requirements.txt
# Requirements for background removal segmentation project
torch==2.0.0
torchvision==0.15.0
streamlit==1.32.0
pillow==9.0.0
numpy==1.23.0
imageio==2.31.0
opencv-python==4.8.0
tqdm==4.66.0
kagglehub==0.3.0